+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 ## unreleased
 
-- add `trim!` function to remove precomputed grid components for lightweight storage
+- add `trim!`/`trim` function to remove precomputed grid components for lightweight storage
 
 ## [1.13.2] - 2025-06-22
 - Allow Triangulate v3